+++ hash.h
@@ -58,7 +58,7 @@ static inline unsigned long hash_long(unsigned long val, unsigned int bits)
 	
 static inline unsigned long hash_ptr(void *ptr, unsigned int bits)
 {
-	return hash_long((unsigned long)ptr, bits);
+	return hash_long((uintptr_t)ptr, bits);
 }
 
 /*
